fix(robloxRegistry): simplify resolves and for loop conditions
My head was cloudy when I first wrote this, the logic was not smooth.
It now properly loops only through the values to set with their supposed conditions on filtering.

This is equivalent to a refactor. But this logic might fix bugs to prior, so it had to be a fix.